Sub tags

+++ src/utils/tag.js
@@ -1,8 +1,9 @@
 
 
 export default {
-  tagOptions: tag => ({dispatch, getters, tasks}) => {
+  tagOptions: tag => ({dispatch, getters, tags, tasks, commit}) => {
     const l = getters['l']
+    const showDelete = !tags.some(el => el.parent === tag.id)
     const opt = [
       {
         name: l['Edit tag'],
@@ -10,6 +11,33 @@ export default {
         callback: () => dispatch('pushPopup', {
             comp: 'AddTag', payload: {...tag, editing: true}, naked: true
           })
+      },
+      {
+        name: l['Add subtag'],
+        icon: 'arrow',
+        callback: () => {
+          dispatch('pushPopup', {comp: 'AddTag', payload: {
+            parent: tag.id,
+          }, naked: true})
+        },
+      },
+      {
+        name: l['Move tag below'],
+        icon: 'tag',
+        callback: () => {
+          dispatch('pushPopup', {
+            comp: 'FastSearch',
+            payload: {
+              callback: (route, {id}) => {
+                if (id !== tag.id)
+                  dispatch('tag/moveTagBelow', {
+                    target: tag.id, tagId: id,
+                  })
+              },
+              allowed: ['tags'],
+            }
+          })
+        },
       },
       {
         name: l["Toggle favorite"],
@@ -31,12 +59,13 @@ export default {
           naked: true
         })
       })
-    opt.push(      {
-      name: l['Delete tag'],
-      icon: 'trash',
-      important: true,
-      callback: () => dispatch('tag/deleteTag', {id: tag.id, tasks})
-    })
+    if (showDelete)
+      opt.push({
+        name: l['Delete tag'],
+        icon: 'trash',
+        important: true,
+        callback: () => dispatch('tag/deleteTag', {id: tag.id, tasks})
+      })
     return opt
   }
 }
